pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
homepage = "https://github.com/radiantlogicinc/fastworkflow"
repository = "https://github.com/radiantlogicinc/fastworkflow"
# documentation = "https://fastworkflow.ai/readthedocs"

[tool.poetry]
name = "fastworkflow"
version = "2.5.0"
description = "A framework for rapidly building large-scale, deterministic, interactive workflows with a fault-tolerant, conversational UX"
authors = ["Dhar Rawal <drawal@radiantlogic.com>"]
license = "Apache-2.0"
readme = "README.md"
keywords = ["fastworkflow", "ai", "workflow", "llm", "openai"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.9.2"
dspy-ai = "2.5.36"
speedict = "^0.3.12"
python-dotenv = "^1.0.1"
scikit-learn = "^1.6.1"
transformers = "^4.48.2"
mmh3 = "^5.1.0"
python-levenshtein = "^0.27.1"
colorama = "^0.4.6"
torch = "^2.7.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
pylint = "^3.3.1"
mypy = "^1.13.0"
bandit = "^1.7.10"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
requests-mock = "^1.12.1"

# poetry install --with aws if you are using AWS Bedrock for LLM or LLM_AGENT
pytest-cov = "^6.2.1"
[tool.poetry.group.aws.dependencies]
boto3 = {version = "^1.38.22", optional = true}

